% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_extract.R
\name{batch_extract}
\alias{batch_extract}
\title{batch_extract}
\usage{
batch_extract(in_dir, write_results = TRUE, out_dir = NULL)
}
\arguments{
\item{in_dir}{Path to directory where thermal images are stored.}

\item{out_dir}{Path to directory where output Rdata file will be stored.}
}
\value{
A list of length equal to the number of thermal images. Each element is a numeric matrix of the
raw infrared data. Also returns a dataframe of parameters unique to each camera.
}
\description{
Batch extraction of raw data from FLIR thermal images.
}
\examples{
results <- batch_extract(system.file("extdata", package = "PatchStatsFLIR"), write_results = FALSE)
}
