% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{connectivity}
\alias{connectivity}
\title{connectivity}
\usage{
connectivity(val_mat, threshold = 1.5)
}
\arguments{
\item{val_mat}{A numeric matrix.}

\item{threshold}{Climate threshold to use for calculation of climate
connectivity (i.e. the amount of change that organisms would be seeking
to avoid).}
}
\value{
A dataframe (one row for each pixel) containing:
 \item{x,y}{The original spatial location of the pixel}
 \item{val}{The pixel value}
 \item{pixel}{The unique id given to the pixel}
 \item{dest_pixel}{The id of the final destination pixel}
 \item{dest_val}{The value of final destination pixel}
 \item{inter_pixel}{The intermediate pixels traversed from origin to
 destination pixel}
 \item{diff_potential}{The potential for change achieved by following
 gradient from hotter to cooler pixels}
 \item{clim_conn}{Climate connectivity, calculated as the maximum potential
 change ('diff_potential') minus the climate threshold. Where this value is
 positive, connectivity from the starting pixel is sufficient to avoid the
 specified threshold of climate warming. See details.}
 @details This measure of climate connectivity and potential for temperature
 change is analogous to that described in
 \href{https://doi.org/10.1073/pnas.1602817113}{McGuire et al. 2016}. The
 basic premise is there is a degree of change that organisms are seeking to
 avoid, and where pixels are sufficiently heterogenous and well connected
 simply moving through the pixels can allow organisms to avoid delterious
 change.
 @references
 McGuire, J. L., Lawler, J. J., McRae, B. H., Nuñez, T. A. and
 Theobald, D. M. (2016), Achieving climate connectivity in a fragmented
 landscape. PNAS, 113: 7195-7200.
\url{https://doi.org/10.1073/pnas.1602817113}
}
\description{
Calculate potential for valerature change and climate connectivity.
}
\examples{

# FLIR valerature matrix ---------------------------------------------------

# Define individual matrix
val_mat <- flir11835$flir_matrix

# Get connectivity
mat_conn <-
connectivity(val_mat = val_mat,
             threshold = 1.5)
head(mat_conn)

}
