% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_edges.R
\name{count_edges}
\alias{count_edges}
\title{count_edges}
\usage{
count_edges(mat, class, bidirectional = FALSE, diagonals = FALSE)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{bidirectional}{Should both sides of each shared edge be counted?
Defaults to FALSE.}

\item{diagonals}{Should diagonal neighbours be included? Defaults to FALSE.}

\item{classes}{The patch class of interest. Must be numeric.}
}
\value{
A numeric value for the number of edges shared by cells of the given
class.
}
\description{
Count the number of edges shared by cells of the same class within a matrix
}
\examples{
# Create matrix
set.seed(317)
cols <- 100
rows <- 50
mat <- matrix(sample(1:3, cols * rows, replace = TRUE),
              nrow = rows, ncol = cols)

# Count edges in each of the three classes (1, 2 or 3)
results <- lapply(1:3, count_edges, mat = mat)
do.call("rbind", results)
[,1]
[1,] 1051
[2,] 1179
[3,] 1028

}
\keyword{internal}
