% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_img.R
\name{crop_img}
\alias{crop_img}
\title{crop_img}
\usage{
crop_img(img, crop_poly = NULL, rtn_format = "matrix")
}
\arguments{
\item{img}{A numeric temperature matrix (such as that returned from
\code{Thermimage::}\code{\link[Thermimage]{raw2temp}}) or raster.}

\item{crop_poly}{If polygon of cropping area is already defined, supply here.
Defaults to NULL.}

\item{rtn_format}{Desired format ('matrix' or 'raster') of the area returned
after cropping. Defaults to 'matrix'.}
}
\value{
A cropped version of the input image.
}
\description{
Crops temperature data to the desired area.
}
\details{
Unless the cropping polygon is already available, function requires 
user input to iteratively refine the size, location and angle (optional) of 
the cropping area.
}
\examples{
# Crop thermal image of wasp nest to the nest only
# N.B. For this example, a good approximation is an ellipse with
# origin coordinates at x = 350, y = 290, x-axis radius of 145
# and y-axis radius of 190.

cropped <- crop_img(img = waspnest_mat)

# Get patches
crop_patches <- get_patches(cropped)

# Plot with patches
plot_patches(df = crop_patches$df,
             patches = crop_patches$patches,
             plot_distribution = FALSE)

}
