% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_mat.R
\name{crop_mat}
\alias{crop_mat}
\title{crop_mat}
\usage{
crop_mat(val_mat)
}
\arguments{
\item{val_mat}{A numeric matrix, such as that returned from
\code{Thermimage::}\code{\link[Thermimage]{raw2temp}}.}
}
\description{
Crops temperature matrix to the desired area.
}
\examples{
path <- "C:\\\\Users\\\\Rebecca\\\\Google Drive\\\\PhD\\\\Ch3\\\\data\\\\flir_sample\\\\Test.jpg"
img <- readflirJPG(path)
settings <-
flirsettings(path)$Info \%>\%
dplyr::bind_rows()
val_mat <- raw2temp(raw = img,
                    E = settings$Emissivity,
                    OD = settings$ObjectDistance,
                    RTemp = settings$ReflectedApparentTemperature,
                    ATemp = settings$AtmosphericTemperature,
                    IRWTemp = settings$IRWindowTemperature,
                    IRT = settings$IRWindowTransmission,
                    RH = settings$RelativeHumidity,
                    PR1 = settings$PlanckR1,
                    PB = settings$PlanckB,
                    PF = settings$PlanckF,
                    PO = settings$PlanckO,
                    PR2 = settings$PlanckR2)

# 350, 290, 145,190
cropped <- crop_mat(val_mat = val_mat)
crop_patches <- get_patches(cropped)
plot_patches(df = crop_patches$df,
patches = crop_patches$patches)

}
