% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flir_metadata.R
\docType{data}
\name{flir_metadata}
\alias{flir_metadata}
\title{Metadata associated with FLIR E40 thermal images}
\format{
A dataframe of environmental parameters with one row for each photo.
The photo number corresponds to the element name in \code{\link{flir_raw}$raw_dat}.
}
\usage{
flir_metadata
}
\description{
Metadata with environmental parameters necessary to accurately convert the
raw data of FLIR thermal images into temperature values.
}
\examples{
flir_metadata
}
\keyword{datasets}
