% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flir_patches.R
\name{flir_patches}
\alias{flir_patches}
\title{get_patches}
\usage{
flir_patches(flir_matrix, photo_no, k = 8, style = "W",
  stats = c("n.patches", "total.area", "obs.edges", "max.edges",
  "max.patch.temp", "min.patch.temp"))
}
\arguments{
\item{flir_matrix}{A matrix returned from the Thermimage package.}

\item{photo_no}{The photo number of the FLIR jpeg from which the matrix was derived.}

\item{k}{Number of neighbours to use when calculating nearest neighbours using spdep::knearneigh}

\item{style}{Style to use when calculating neighbourhood weights using spdep::nb2listw.}

\item{stats}{Patch statistics to return from SDMTools::ClassStat.}
}
\value{
List including dataframe of temperature data with pixels matched to patches, patches as a SpatialPolygonsDataFrame, and a dataframe with overall patch statistics.
}
\description{
Find hot and cold spots in a FLIR thermal image, and calculate their patch statistics.
}
\examples{
data("FLIR11835")
flir_patches(flir_matrix = FLIR1185, photo_no = FLIR_no)
}
