% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flir_stack.R
\docType{data}
\name{flir_stack}
\alias{flir_stack}
\title{Raster stack of FLIR images}
\format{A RasterStack}
\usage{
flir_stack
}
\description{
Results of running \code{stack_imgs} for FLIR images taken at different times
of the day.
}
\examples{
raster::plot(flir_stack)
}
\keyword{datasets}
