% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R
\name{kurtosis}
\alias{kurtosis}
\title{kurtosis}
\usage{
kurtosis(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}
}
\description{
Helper function to calculate kurtosis, using
\code{moments::}\code{\link[moments]{kurtosis}}.
}
\examples{
kurtosis(1:100)
kurtosis(matrix(1:100))
}
\keyword{internal}
