% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_sapply.R
\name{multi_sapply}
\alias{multi_sapply}
\title{multi_sapply}
\usage{
multi_sapply(...)
}
\arguments{
\item{...}{First argument should be the vector, all subsequent arguments
are functions to apply (naming functions is optional).}
}
\value{
A wide dataframe with a column for the results of each function.
}
\description{
Apply multiple functions to a numeric vector.
}
\details{
NA values are always removed.
}
\references{
Code snippet by \href{http://bogumilkaminski.pl/about/}{Bogumił Kamiński},
adapted from:
\url{http://rsnippets.blogspot.co.uk/2011/11/applying-multiple-functions-to-data.html}
}
\keyword{internal}
