% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_stats.R
\name{patch_stats}
\alias{patch_stats}
\title{patch_stats}
\usage{
patch_stats(mat, class)
}
\arguments{
\item{mat}{The matrix.}

\item{class}{The patch class for which statistics should be calculated.}
}
\description{
Calculate patch statistics for patch classes of interest in a numeric matrix.
}
\examples{
# Create matrix
matrix(c(1,1,2,1,
         1,0,2,0,
         1,2,0,1,
         2,2,0,1), nrow = 4, ncol  = 4,
         byrow = TRUE)

# Count edges in each of the three classes (1, 2 or 3)
count_edges(mat)
}
