% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_5.R
\name{perc_5}
\alias{perc_5}
\title{perc_5}
\usage{
perc_5(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector over which to calculate 5th percentile.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}
}
\description{
Helper function to calculate 5th percentile.
}
\examples{
perc_5(c(1:100, NA))
}
\keyword{internal}
