% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_95.R
\name{perc_95}
\alias{perc_95}
\title{perc_95}
\usage{
perc_95(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector over which to calculate 95th percentile.}

\item{na.rm}{Whether to remove NAs. Defaults to TRUE.}
}
\value{
A named vector.
}
\description{
Helper function to calculate 95th percentile.
}
\examples{
perc_95(c(1:100, NA))
}
