% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patches.R
\name{plot_patches}
\alias{plot_patches}
\title{plot_patches}
\usage{
plot_patches(df, patches, plot_distribution = TRUE, print_plot = FALSE,
  save_plot = TRUE, return_plot = FALSE, out_dir, file_name = NULL,
  file_ext = "png", lab_size = 8, text_size = 6, fig_width = 8,
  fig_height = 9, fig_units = "cm", val_pal = c("black", "#050155",
  "#120172", "#3b008e", "#7200a9", "#8f00a0", "#ba187f", "#d9365b", "#ed5930",
  "#f76323", "#fa8600", "#f6a704", "#fad61e", "#fad61e"),
  patch_cols = c("mistyrose", "cornflowerblue"), patch_labs = c("Hot spots",
  "Cold spots"))
}
\arguments{
\item{df}{A dataframe returned from \code{get_patches}.}

\item{patches}{A SpatialPolygonsDataFrame returned from \code{get_patches}.}

\item{print_plot}{Should the resulting plots be printed? Defaults to FALSE.}

\item{save_plot}{Should the resulting plots be saved? Defaults to TRUE.}

\item{out_dir}{Path to directory where plots will be saved.}

\item{lab_size}{Size of axes labels. Defaults to 8.}

\item{text_size}{Size of axes text and legend text. Defaults to 6.}

\item{val_pal}{Colour palette to use for raster. Defaults to palette derived from a FLIR jpeg.}

\item{patch_cols}{Colours for the patch borders (hot spot colour followed by cold spot colour).}
}
\description{
Plot hot and cold spots.
}
\examples{
# Find hot and cold spots
results <- get_patches(flir_matrix = flir11835$flir_matrix,matrix_id = flir11835$matrix_id)

# Plot
df <- results$df
patches <- results$patches
plot_patches(df = df, patches = patches, print_plot = TRUE, save_plot = FALSE)

}
