% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stack.R
\name{plot_stack}
\alias{plot_stack}
\title{plot_stack}
\usage{
plot_stack(
  img_stack,
  n = 100,
  print_plot = TRUE,
  save_plot = FALSE,
  return_plot = FALSE,
  out_dir = NULL,
  file_name = NULL,
  file_ext = "png",
  fig_width = 8,
  fig_height = 9,
  fig_units = "cm",
  xlabel = NULL,
  ylabel = NULL,
  lab_size = 8,
  text_size = 6,
  point_size = 0.5,
  quant_linetype = "dotted",
  quant_colour = "black",
  y_breaks = waiver(),
  custom_theme = NULL
)
}
\arguments{
\item{img_stack}{The RasterStack to plot.}

\item{n}{The number of pixels to sample from each raster. Defaults to 100.}

\item{print_plot}{Should the resulting plots be printed? Defaults to FALSE.}

\item{save_plot}{Should the resulting plots be saved? Defaults to TRUE.}

\item{return_plot}{Should the resulting plots be returned? Defaults to FALSE.}

\item{out_dir}{Path to directory where plots should be saved (if applicable).}

\item{file_name}{Plot filename (if applicable).}

\item{file_ext}{File extension. Defaults to '.png'.}

\item{fig_width}{Figure width (if saved). Defaults to 8.}

\item{fig_height}{Figure height (if saved). Defaults to 9.}

\item{fig_units}{Figure dimension units (if saved). Defaults to "cm".}

\item{xlabel}{Option to specify the x axis label.}

\item{ylabel}{Option to specify the y axis label.}

\item{lab_size}{Size of axis labels. Defaults to 8.}

\item{text_size}{Size of axis text. Defaults to 6.}

\item{point_size}{Point size for raw data. Defaults to 0.5}

\item{quant_linetype}{Linetype for the upper and lower quantile lines. 
Defaults to "dotted".}

\item{quant_colour}{Line colour for the upper and lower quantile lines. 
Defaults to "black".}

\item{y_breaks}{Option to manually specify breaks in y axis. Defaults
to \code{waiver()}, where breaks are computed by the transformation object
(see \code{ggplot2::}\code{scale_colour_gradient}).}

\item{custom_theme}{Option to replace existing theme with a customised 
\code{ggplot} theme. See \code{ggplot2::}\code{theme}.}
}
\description{
Create violin plots of raster stacks, such as those returned from 
\code{\link{stack_imgs}}. Useful to assess the distribution of data across
replicates e.g. over time.
}
\examples{
plot_stack(img_stack = flir_stack, 
           n = 100,
           print_plot = TRUE,
           save_plot = FALSE,
           xlabel = "Time",
           ylabel = "Temperature",
           y_breaks = seq(25, 35, 1))
           
\dontrun{
# Specify custom theme
library(ggplot2)
custom_theme <- 
    theme_classic() + 
    theme(axis.title = element_text(size = 8),
          axis.text.x = element_text(size = 7,
          angle = 45,
          vjust = 0.5),
          axis.text.y = element_text(size = 7))
          
plot_stack(img_stack = flir_stack,
           print_plot = TRUE,
           save_plot = FALSE,
           custom_theme = custom_theme)
}
}
