% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHDI.R
\name{SHDI}
\alias{SHDI}
\title{SHDI}
\usage{
SHDI(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Whether to remove NAs. Defaults to TRUE.}
}
\value{
A named vector.
}
\description{
Helper function to calculate Shannon Diversity Index.
}
\examples{
SHDI(c(1:100, NA))
}
