% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIDI.R
\name{SIDI}
\alias{SIDI}
\title{SIDI}
\usage{
SIDI(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}
}
\description{
Helper function to calculate Simpson Diversity Index.
}
\examples{
SIDI(c(1:100, NA))
}
\keyword{internal}
