% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{skewness}
\usage{
skewness(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}
}
\description{
Helper function to calculate skewness, using
\code{moments::}\code{\link[moments]{skewness}}.
}
\keyword{internal}
