% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_img.R
\name{stack_imgs}
\alias{stack_imgs}
\title{stack_imgs}
\usage{
stack_imgs(img_list)
}
\arguments{
\item{img_list}{List of temperature matrices to convert.}
}
\value{
RasterStack with one layer for each list element.
}
\description{
Helper function to create raster stack from a list of matrices.
}
\examples{
# Load raw data
raw_dat <- flir_raw$raw_dat
camera_params <- flir_raw$camera_params
metadata <- flir_metadata

# Batch convert
img_list <- batch_convert(raw_dat, write_results = FALSE)

# Stack
img_stack <- stack_imgs(img_list)
raster::plot(img_stack)
}
