% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_by_group.R
\name{stats_by_group}
\alias{stats_by_group}
\title{stats_by_group}
\usage{
stats_by_group(metadata, mat_list, matrix_ID = "photo_no", subset_var,
  subset_val, round_val = 1, ...)
}
\arguments{
\item{metadata}{A dataframe denoting the grouping of different matrices.}

\item{mat_list}{List of temperatures matrices.}

\item{matrix_ID}{Name of the metadata variable that identifies unique temperature matrices. Should match element names in the list of matrices.}

\item{subset_var}{The name of the metadata variable that denotes the grouping of temperature matrices.}

\item{subset_val}{The value of the grouping variable denoted by `subset_var` -- designed for use in conjunction with `sapply`.}

\item{round_val}{Value to round to. Defaults to 1 (i.e. no rounding).}
}
\value{
A list containing:
 \item{raw_dat}{A list with one element per input thermal image. Each element is a numeric matrix of the
raw infrared data.}
 \item{camera_params}{A dataframe of callibration constants unique to each camera.}
}
\description{
Calculate thermal statistics across photos within groups.
}
\examples{
# Load raw data
raw_dat <- flir_raw$raw_dat
camera_params <- flir_raw$camera_params
metadata <- flir_raw$metadata

# Batch convert
mat_list <- batch_convert(raw_dat, write_results = FALSE)

# Stats

}
