% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sulawesi_bg.R
\docType{data}
\name{sulawesi_bg}
\alias{sulawesi_bg}
\title{Boundary polygon of Sulawesi}
\format{A SpatialPolygonsDataFrame}
\usage{
sulawesi_bg
}
\description{
Boundary polygon of Sulawesi for plotting with Worldclim2 raster data 
\code{sulawesi_temp}.
}
\examples{
sp::plot(sulawesi_bg)
}
\keyword{datasets}
