% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim_data.R
\docType{data}
\encoding{UTF-8}
\name{sulawesi_temp}
\alias{sulawesi_temp}
\title{WorldClim2 mean monthly temperature data for Sulawesi}
\format{
A geographic temperature raster
}
\usage{
sulawesi_temp
}
\description{
Gridded temperature data at 30 arc-second resolution for the island of
Sulawesi. Temperature is average January temperature (°C) for the 'current'
time period (1970-2000).
}
\examples{
raster::plot(sulawesi_temp)
}
\references{
Fick, S.E. and Hijmans, R.J. (2017),
Worldclim 2: New 1-km spatial resolution climate surfaces for global land
areas. International Journal of Climatology. Available at:
\url{http://worldclim.org/version2}.
}
\keyword{datasets}
