% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermStats.R
\docType{package}
\name{ThermStats}
\alias{ThermStats}
\alias{ThermStats-package}
\title{\code{ThermStats} package}
\description{
Calculate statistics for thermal images
}
\details{
See the README on
\href{https://github.com/rasenior/ThermStats#readme}{GitHub}
}
