% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waspnest_mat.R
\docType{data}
\name{waspnest_mat}
\alias{waspnest_mat}
\title{Temperature matrix from thermal image of wasp nest}
\format{
A 640 x 480 numeric temperature matrix.
}
\usage{
waspnest_mat
}
\description{
The data was extracted from a FLIR SC660 thermal image of a wasp nest,
using the package \pkg{Thermimage}. An alternative is
to use the freely availble software
\href{http://www.flir.com/instruments/display/?id=54865}{FLIR Tools}, with
which this image is freely distributed.
}
\examples{
mean(waspnest_mat)
hist(waspnest_mat)
# Convert to raster
waspnest_raster <- raster::raster(waspnest_mat)
# Visualise using the rasterVis package
rasterVis::levelplot(waspnest_raster,
                     layers = 1,
                     margin = list(FUN = 'median'),
                     contour= FALSE)
}
\references{
\pkg{Thermimage}, available on
\href{https://cran.r-project.org/package=Thermimage}{CRAN} and
\href{https://github.com/gtatters/Thermimage}{GitHub}

\href{http://www.flir.com/instruments/display/?id=54865}{FLIR Tools}
}
\keyword{datasets}
