% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_by_group.R
\name{average_by_group}
\alias{average_by_group}
\title{average_by_group}
\usage{
average_by_group(metadata, mat_list, matrix_id, grouping_var, round_val)
}
\arguments{
\item{metadata}{A dataframe denoting the grouping of different matrices.}

\item{mat_list}{List of matrices.}

\item{matrix_id}{Name of the metadata variable that identifies unique
matrices. Should match element names in the list of matrices.}

\item{grouping_var}{The name of the metadata variable that denotes the
grouping of matrices.}

\item{round_val}{Value to round to. Use 1 for no rounding.}

\item{k}{Number of neighbours to use when calculating nearest neighbours
using \code{spdep::}\code{\link[spdep]{knearneigh}}.}

\item{style}{Style to use when calculating neighbourhood weights using
\code{spdep::}\code{\link[spdep]{nb2listw}}.}

\item{...}{Use to specify summary statistics that should be calculated across
all pixels. Several helper functions are included for use here:
\code{\link{perc_5}}, \code{\link{perc_95}},
\code{\link{SHDI}}, \code{\link{SIDI}},
\code{\link{kurtosis}} and \code{\link{skewness}}. Also see examples below.}
}
\value{
A dataframe with pixel and patch statistics. See
\code{\link{patch_stats}} for details of all the patch statistics returned.
}
\description{
Average multiple matrices or rasters within groups.
}
