% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_by_group.R
\name{average_by_group}
\alias{average_by_group}
\title{average_by_group}
\usage{
average_by_group(metadata, img_list, id, grouping_var, round_val = NULL)
}
\arguments{
\item{metadata}{A dataframe denoting the grouping of different images.}

\item{img_list}{List or stack of numeric temperature matrices or rasters.}

\item{id}{Name of the metadata variable that identifies unique
images. Should match element names in the image list.}

\item{grouping_var}{The name of the metadata variable that denotes the
grouping of images.}

\item{round_val}{Number of decimals to round to (optional).}
}
\value{
If \code{img_list} is supplied as a list of matrices, returns
a new list of matrices with one element for each unique value of the grouping
variable. 

If \code{img_list} is supplied as a raster stack, returns a new raster
stack with one raster for each unique value of the grouping variable. 

Grouping variables with no matching images are dropped.
}
\description{
Average multiple matrices or rasters within groups.
}
\examples{
# Load raw data
raw_dat <- flir_raw$raw_dat
camera_params <- flir_raw$camera_params
metadata <- flir_metadata

# Batch convert
img_list <- batch_convert(raw_dat, write_results = FALSE)

# With list of matrices -----------------------------------------------------

# Average matrices in list
avg_list <- average_by_group(metadata = metadata,
                             img_list = img_list,
                             id = "photo_no",
                             grouping_var = "rep_id")
                             
# With raster stack ---------------------------------------------------------

# Coerce matrix list to raster stack
mat_stack <- stack_imgs(img_list)

# Average in raster stack
avg_stack <- average_by_group(metadata = metadata,
                              img_list = mat_stack,
                              id = "photo_no",
                              grouping_var = "rep_id")
# Plot
raster::plot(avg_stack)

# Compare to invididual pics
# Photos 8565 and 8583 were averaged to make T7P1
# Photos 8589 and 8613 were averaged to make T7P2

allstack <- raster::stack(mat_stack, avg_stack)
rasterVis::levelplot(allstack,
                    layout=c(2, 3),
                    index.cond=list(c(1, 3, 2, 4, 5, 6)))
                    
}
