% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_convert.R
\name{batch_convert}
\alias{batch_convert}
\title{batch_convert}
\usage{
batch_convert(raw_dat, E = 1, OD = 1, RTemp = 20, ATemp = RTemp,
  IRWTemp = RTemp, IRT = 1, RH = 50, PR1 = 21106.77, PB = 1501,
  PF = 1, PO = -7340, PR2 = 0.012545258, write_results = TRUE,
  out_dir = NULL)
}
\arguments{
\item{raw_dat}{A list of raw FLIR matrices returned from \code{batch_extract}.}

\item{out_dir}{Path to directory where converted temperature data will be written.}
}
\description{
Batch convert list of raw FLIR matrices.
}
\examples{
# Load raw data
raw_dat <- flir_raw$raw_dat
camera_params <- flir_raw$camera_params
metadata <- flir_raw$metadata

# Create vector denoting the position of each element of raw_dat (each photo)
# within the metadata dataframe
photo_index <- match(names(raw_dat), metadata$photo_no)

# Define each of the environmental parameters

# Emissivity is mean of range given in Scheffers et al. 2016
E=mean(c(0.982,0.99))
# Object distance is calculated as the hypotenuse of a right
# triangle where the vertical side is 1.3 m (breast height)
# and the angle down is 45 degrees
OD=(sqrt(2))*1.3
# Apparent reflected temperature of the environment is set as
# dry bulb temperature from the hygrometer.
RTemp=metadata$atm_temp[photo_index]
# Atmospheric temperature is the same as reflected temperature
# (default).
ATemp=metadata$atm_temp[photo_index]
# Infrared Window Temperature is the same as reflected temperature
# (default).
IRWTemp=metadata$atm_temp[photo_index]
# Infrared Window transmission is kept at default.
IRT=1
# Relative humidity is set as humidity from the hygrometer.
RH=metadata$rel_humidity[photo_index]

# Camera parameters are constant for a given camera
PR1=camera_params[,"PlanckR1"]
PB=camera_params[,"PlanckB"]
PF=camera_params[,"PlanckF"]
PO=camera_params[,"PlanckO"]
PR2=camera_params[,"PlanckR2"]

results <- batch_convert(raw_dat, E, OD, RTemp, ATemp, IRWTemp, IRT, RH, PR1, PB, PF, PO, PR2, write_results = FALSE)
}
