% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_extract.R
\name{batch_extract}
\alias{batch_extract}
\title{batch_extract}
\usage{
batch_extract(in_dir, write_results = TRUE, out_dir = NULL,
  file_name = NULL, inc = NULL, exc = NULL)
}
\arguments{
\item{in_dir}{Path to directory where thermal images are stored.}

\item{write_results}{Should the results be written as an Rdata file?
Defaults to true.}

\item{out_dir}{Path to directory where output Rdata file will be stored.
Defaults to working directory.}

\item{file_name}{File name (without extension). Defaults to NULL, with the 
name 'flir_raw_' plus the current date. File format is .Rds}

\item{inc}{Vector of photo numbers to include. Defaults to NULL, meaning 
all files are included.}

\item{exc}{Vector of photo numbers to exclude. Defaults to NULL, meaning 
all files are included.}
}
\value{
A list containing:
 \item{raw_dat}{A list with one element per input thermal image. Each element
 is a numeric matrix of the raw infrared data.}
 \item{camera_params}{A dataframe of callibration constants unique to each
 camera.}
}
\description{
Batch extraction of raw data from FLIR thermal images.
}
\details{
Batch implementation of \code{Thermimage::}\code{\link[Thermimage]{readflirJPG}}.
}
\examples{
# Batch extract four FLIR thermal images included in this package.
results <-
    batch_extract(system.file("extdata", package = "ThermStats"),
                  write_results = FALSE)
}
