% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_extract.R
\name{batch_extract}
\alias{batch_extract}
\title{batch_extract}
\usage{
batch_extract(in_dir, write_results = TRUE, out_dir = NULL,
  file_name = NULL, inc = NULL, exc = NULL,
  exiftoolpath = "installed")
}
\arguments{
\item{in_dir}{Path to directory where thermal images are stored.}

\item{write_results}{Should the results be written as an Rdata file?
Defaults to true.}

\item{out_dir}{Path to directory where output Rdata file will be stored.
Defaults to working directory.}

\item{file_name}{File name (without extension). Defaults to NULL, with the 
name 'flir_raw_' plus the current date. File format is .Rds}

\item{inc}{Vector of photo numbers to include. Defaults to NULL, meaning 
all files are included.}

\item{exc}{Vector of photo numbers to exclude. Defaults to NULL, meaning 
all files are included.}

\item{exiftoolpath}{Passed directly to 
\code{Thermimage::}\code{\link[Thermimage]{readflirJPG}}: 
"A character string that determines whether Exiftool has been 'installed' 
(http://www.sno.phy.queensu.ca/~phil/exiftool/) or not. If Exiftool has been 
installed in a specific location, use to direct to the folder location."}
}
\value{
A list containing:
 \item{raw_dat}{A list with one element per input thermal image. Each element
 is a numeric matrix of the raw infrared data.}
 \item{camera_params}{A dataframe of callibration constants unique to each
 camera.}
}
\description{
Batch extraction of raw data from FLIR thermal images.
}
\details{
Batch implementation of \code{Thermimage::}\code{\link[Thermimage]{readflirJPG}}.

Note that Exiftool is essential for this function to work correctly. See the
documentation of \code{Thermimage::}\code{\link[Thermimage]{readflirJPG}}:

"Exiftool should install on most operating systems. Consult with 
http://www.sno.phy.queensu.ca/~phil/exiftool/ for information on installing 
Exiftool. If trouble installing, download Exiftool and set exiftoolpath to 
the custom folder location. To test if the custom path to Exiftool will work 
on your OS, try your own system or system2 call: 
system2('/custompath/exiftool') to see if you get an error or not."
}
\examples{
# Batch extract four FLIR thermal images included in this package.
results <-
    batch_extract(system.file("extdata", package = "ThermStats"),
                  write_results = FALSE)
}
\references{
Exiftool Command line tool: http://www.sno.phy.queensu.ca/~phil/exiftool/
}
