% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{connectivity}
\alias{connectivity}
\title{connectivity}
\usage{
connectivity(mat, conn_thresh = 1.5)
}
\arguments{
\item{mat}{A numeric temperature matrix.}

\item{conn_thresh}{Climate threshold to use for calculation of thermal
connectivity (i.e. the amount of change that organisms would be seeking
to avoid).}
}
\value{
A dataframe (one row for each pixel) containing:
 \item{x,y}{The original spatial location of the pixel}
 \item{val}{The pixel value}
 \item{pixel}{The unique id given to the pixel}
 \item{dest_pixel}{The id of the final destination pixel}
 \item{dest_val}{The value of final destination pixel}
 \item{inter_pixel}{The intermediate pixels traversed from origin to
 destination pixel}
 \item{diff_potential}{The potential for change achieved by following
 gradient from hotter to cooler pixels}
 \item{therm_conn}{Thermal connectivity, calculated as the maximum potential
 change (\code{diff_potential}) minus \code{conn_thresh}. Where
 this value is positive, connectivity of the starting pixel is sufficient to
 avoid the specified threshold of warming. See details.}
}
\description{
Calculate thermal connectivity and potential for temperature change.
}
\details{
This measure of climate connectivity and potential for temperature
change is analogous to that described in
\href{https://doi.org/10.1073/pnas.1602817113}{McGuire et al. 2016}. The
basic premise is that there is a degree of change that organisms are seeking
to avoid, and when pixels are sufficiently heterogenous and well connected
organisms can move through the pixels to avoid delterious temperature change.
}
\examples{
# Define matrix as FLIR thermal image
mat <- flir11835$flir_matrix

# Get connectivity
mat_conn <-
connectivity(mat = mat,
             conn_thresh = 1.5)
head(mat_conn)

# Plot the potential for temperature change
library(ggplot2)
ggplot(mat_conn, 
       aes(x = x, y = y, fill = diff_potential))+
    geom_raster() +
    scale_fill_viridis_c()
               
}
\references{
McGuire, J. L., Lawler, J. J., McRae, B. H., Nunez, T. A. and
Theobald, D. M. (2016), Achieving climate connectivity in a fragmented
landscape. PNAS, 113: 7195-7200.
\url{https://doi.org/10.1073/pnas.1602817113}
}
