% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_edges.R
\name{count_edges}
\alias{count_edges}
\title{count_edges}
\usage{
count_edges(mat, class, bidirectional = FALSE, diagonals = FALSE)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{bidirectional}{Should both sides of each shared edge be counted?
Defaults to FALSE.}

\item{diagonals}{Should diagonal neighbours be included? Defaults to FALSE.}

\item{classes}{The patch class of interest. Must be numeric.}
}
\value{
A numeric value for the number of edges shared by cells of the given
class.
}
\description{
Count the number of edges shared by cells of the same class within a matrix
}
\keyword{internal}
