% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_edges.R
\name{count_edges}
\alias{count_edges}
\title{count_edges}
\usage{
count_edges(mat, classes = NULL)
}
\arguments{
\item{in_dir}{Path to directory where thermal images are stored.}

\item{write_results}{Should the results be written as an Rdata file? Defaults to true.}

\item{out_dir}{Path to directory where output Rdata file will be stored. Defaults to working directory.}
}
\value{
A list containing:
 \item{raw_dat}{A list with one element per input thermal image. Each element is a numeric matrix of the
raw infrared data.}
 \item{camera_params}{A dataframe of callibration constants unique to each camera.}
}
\description{
Count the number of edges shared by cells of the same class within a matrix
}
\examples{
# Batch extract four FLIR thermal images included in this package.
results <- batch_extract(system.file("extdata", package = "PatchStatsFLIR"), write_results = FALSE)
}
