% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_img.R
\name{crop_img}
\alias{crop_img}
\title{crop_img}
\usage{
crop_img(img)
}
\arguments{
\item{img}{A numeric temperature matrix, such as that returned from
\code{Thermimage::}\code{\link[Thermimage]{raw2temp}}.}
}
\value{
A cropped version of the input matrix.
}
\description{
Crops temperature matrix to the desired area.
}
\details{
Requires user input to iteratively refine the size, location and
angle (optional) of the cropping area.
}
\examples{
# Crop thermal image of wasp nest to the nest only
# N.B. For this example, a good approximation is an ellipse with
# origin coordinates at x = 350, y = 290, x-axis radius of 145
# and y-axis radius of 190.

cropped <- crop_img(img = waspnest_mat)

# Get patches
crop_patches <- get_patches(cropped)

# Plot with patches
plot_patches(df = crop_patches$df,
             patches = crop_patches$patches,
             plot_distribution = FALSE)

}
