% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\docType{data}
\name{flir_raw}
\alias{flir_raw}
\title{FLIR E40 thermal image data}
\format{A list consisting of:
  \itemize{
    \item{flir_matrix} {= A 120 x 160 numeric temperature matrix.}
    \item{photo_no} {= The photo number.}
  }}
\usage{
flir_raw
}
\description{
The data was extracted from a FLIR E40 thermal image using
the package \code{Thermimage}. An alternative is
to use the freely availble software 'FLIR Tools'
(\url{http://www.flir.com/instruments/display/?id=54865}).
}
\examples{
hist(flir_raw$raw_list$`8565`)
}
\references{
\code{Thermimage}, available on CRAN
  (\url{https://cran.r-project.org/package=Thermimage}) and GitHub
  (\url{https://github.com/gtatters/Thermimage}).

  FLIR Tools (\url{http://www.flir.com/instruments/display/?id=54865}).
}
\keyword{datasets}
