% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flir_raw.R, R/raw_data.R
\docType{data}
\name{flir_raw}
\alias{flir_raw}
\title{Raw FLIR E40 thermal image data and camera parameters}
\format{A list consisting of:
  \itemize{
    \item{raw_dat} {= A list of 120 x 160 numeric matrices of the raw 16-bit encoded
    values from four FLIR thermal images included in the package distribution.
    Extracted using \code{Thermimage::}\code{\link[Thermimage]{readflirJPG}}.}
    \item{camera_params} {= A dataframe of calibration constants unique to
    each FLIR camera.
    Extracted using \code{Thermimage::}\code{\link[Thermimage]{flirsettings}}.}
  }}
\usage{
flir_raw

flir_raw
}
\description{
The data was extracted from a FLIR E40 thermal image using
the package \pkg{Thermimage}. An alternative is
to use the freely availble software
\href{http://www.flir.com/instruments/display/?id=54865}{FLIR Tools}.

The data was extracted from a FLIR E40 thermal image using
the package \pkg{Thermimage}. An alternative is
to use the freely availble software
\href{http://www.flir.com/instruments/display/?id=54865}{FLIR Tools}.
}
\examples{
hist(flir_raw$raw_dat$`8565`)
flir_raw$camera_params
hist(flir_raw$raw_dat$`8565`)
}
\references{
\pkg{Thermimage}, available on
\href{https://cran.r-project.org/package=Thermimage}{CRAN} and
\href{https://github.com/gtatters/Thermimage}{GitHub}

\href{http://www.flir.com/instruments/display/?id=54865}{FLIR Tools}

\itemize{
\item \pkg{Thermimage} is available on
  \href{https://cran.r-project.org/package=Thermimage}{CRAN} and
  \href{https://github.com/gtatters/Thermimage}{GitHub}.
\item FLIR Tools is available at \url{http://www.flir.com/instruments/display/?id=54865}.
}
}
\keyword{datasets}
