% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_patches.R
\name{get_patches}
\alias{get_patches}
\title{get_patches}
\usage{
get_patches(flir_matrix, photo_no, k = 8, style = "W",
  stats = c("n.patches", "total.area", "obs.edges", "max.edges",
  "max.patch.temp", "min.patch.temp"))
}
\arguments{
\item{flir_matrix}{A numeric temperature matrix, such as that returned from \code{Thermimage}.}

\item{photo_no}{The photo number of the FLIR jpeg from which the matrix was derived.}

\item{k}{Number of neighbours to use when calculating nearest neighbours using \code{spdep::knearneigh}.}

\item{style}{Style to use when calculating neighbourhood weights using \code{spdep::nb2listw}.}

\item{stats}{Patch statistics to return from \code{SDMTools::ClassStat}.}
}
\value{
A list containing:
 \item{flir_df}{A dataframe with one row for each pixel, and variables denoting:
 the original position of the pixel (y and x); its temperature (temp); its
 Z value from \code{spdep::localG}; its classification (G_bin) into hot (1) or cold
 spots (-1) according to the Z value (see \code{?localG}); the unique ID
 of the patch in which the pixel fell; and the photo number.}
 \item{patches}{A SpatialPolygonsDataFrame of hot and cold spots. Hot spots
 have a value of 1, and cold spots a value of -1.}
 \item{patch_stats}{A dataframe with patch statistics for hot spots and cold spots,
 respectively.}
}
\description{
Find hot and cold spots in a FLIR thermal image, and calculate their patch statistics.
}
\examples{
# Find hot and cold spots
results <- get_patches(flir_matrix = flir11835$flir_matrix,photo_no = flir11835$photo_no)

# Look at the results for individual pixels
head(results$flir_df)

# Look at the patch statistics for hot and cold spots
results$patch_stats

# Plot the patches
require(sp)
plot(results$patches)
}
