% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_stats.R
\name{patch_stats}
\alias{patch_stats}
\title{patch_stats}
\usage{
patch_stats(mat, class)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{class}{The patch class for which statistics should be calculated.}
}
\value{
A dataframe containing:
 \item{class}{The patch class for which statistics were calculated.}
 \item{area}{Total area (number of cells).}
 \item{prop}{Proportion of landscape area represented by cells in the given class.}
 \item{perim}{Total perimeter (number of outside edges).}
 \item{perim_min}{Minimum perimeter, given the number of cells.}
 \item{inside_edges}{Total number of shared, inside edges.}
 \item{inside_edges_max}{Maximum number of inside edges that could be shared,
 given the number of cells.}
 \item{shape_index}{Patch shape index. Calculated as the observed perimeter
 divided by the minimum perimeter.
 See \href{http://www.umass.edu/landeco/research/fragstats/documents/fragstats.help.4.2.pdf}{FRAGSTATS} p.101.}
 \item{AI}{Aggregation index. Calculated as the observed number of shared,
 inside edges divided by the maximum number that could be shared.
 See \href{https://doi.org/10.1023/A:1008102521322}{He et al. 2000}.}
 \item{cohesion}{Cohesion index. Calculated from the area and perimeter of
 cells in the given class relative to the area of the entire landscape.
 See \href{https://doi.org/10.2307/2265590}{Scuhmaker 1996}.}
}
\description{
Calculate patch statistics for a given class (i.e. cell value) in a numeric matrix.
}
\details{
Statistics are based on and inspired by
\code{SDMTools::}\code{\link[SDMTools]{ClassStat}},
\href{http://www.umass.edu/landeco/research/fragstats/fragstats.html}{FRAGSTATS}
and \href{https://doi.org/10.1111/2041-210X.12488}{Faye et al. 2016}
}
\examples{
# Create matrix
mat <- matrix(c(1,1,2,1,
                1,0,2,0,
                1,2,0,1,
                2,2,0,1), nrow = 4, ncol  = 4,
              byrow = TRUE)

patch_stats(mat = mat, class = 1)

}
\references{
\itemize{
\item \pkg{SDMTools}: Species Distribution Modelling Tools: Tools for
processing data associated with species distribution modelling exercises.
Available on \href{https://CRAN.R-project.org/package=SDMTools}{CRAN}.
\item McGarigal, K., S. A. Cushman, M. C. Neel, and E. Ene. 2002.
FRAGSTATS: Spatial Pattern Analysis Program for Categorical Maps. Computer
software program produced by the authors at the University of Massachusetts,
Amherst. Available at the following web site:
\url{www.umass.edu/landeco/research/fragstats/fragstats.html}
\item Faye, E. , Rebaudo, F. , Yánez‐Cajo, D. , Cauvy‐Fraunié, S. ,
Dangles, O. and Tatem, A. (2016), A toolbox for studying thermal
heterogeneity across spatial scales: from unmanned aerial vehicle
imagery to landscape metrics. Methods Ecol Evol, 7: 437-446.
\url{https://doi.org/10.1111/2041-210X.12488}
\item He, H.S., DeZonia, B.E. & Mladenoff, D.J. (2000),
An aggregation index (AI) to quantify spatial patterns of landscapes.
Landscape Ecology, 15: 591.
\url{https://doi.org/10.1023/A:1008102521322}
\item Schumaker, N. H. (1996),
Using landscape indices to predict habitat connectivity.
Ecology, 77: 1210-1225.
\url{https://doi.org/10.2307/2265590}
}
}
