% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_stats.R
\name{patch_stats}
\alias{patch_stats}
\title{patch_stats}
\usage{
patch_stats(val_mat, class)
}
\arguments{
\item{val_mat}{A numeric matrix.}

\item{class}{The patch class for which statistics should be calculated.}
}
\value{
A dataframe containing:
 \item{class}{The patch class for which statistics were calculated.}
 \item{area}{Total area (number of cells).}
 \item{prop}{Proportion of landscape area represented by cells in the given class.}
 \item{perim}{Total perimeter (number of outside edges).}
 \item{perim_min}{Minimum perimeter, given the number of cells.}
 \item{inside_edges}{Total number of shared, inside edges.}
 \item{inside_edges_max}{Maximum number of inside edges that could be shared,
 given the number of cells.}
 \item{shape_index}{Patch shape index. Calculated as the observed perimeter
 divided by the minimum perimeter.}
 \item{AI}{Aggregation index. Calculated as the observed number of shared,
 inside edges divided by the maximum number that could be shared. See He et al.}
 \item{cohesion}{Cohesion index. Calculated from the area and perimeter of
 cells in the given class relative to the area of the entire landscape.
 See Scuhmaker 1996.}
}
\description{
Calculate patch statistics for a given class (i.e. cell value) in a numeric matrix.
}
\details{
Statistics are based on and inspired by [SDMTools::ClassStat()],
[FRAGSTATS](http://www.umass.edu/landeco/research/fragstats/fragstats.html)
and [Faye et al. 2016](https://doi.org/10.1111/2041-210X.12488 ).
}
\examples{
# Create matrix
matrix(c(1,1,2,1,
         1,0,2,0,
         1,2,0,1,
         2,2,0,1), nrow = 4, ncol  = 4,
         byrow = TRUE)

# Count edges in each of the three classes (1, 2 or 3)
count_edges(val_mat)
}
\references{
McGarigal, K., S. A. Cushman, M. C. Neel, and E. Ene. 2002.
FRAGSTATS: Spatial Pattern Analysis Program for Categorical Maps. Computer
software program produced by the authors at the University of Massachusetts,
Amherst. Available at the following web site:
www.umass.edu/landeco/research/fragstats/fragstats.html
}
