% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patches.R
\name{plot_patches}
\alias{plot_patches}
\title{plot_patches}
\usage{
plot_patches(df, patches, bg_poly = NULL, bg_colour = NULL,
  facet_cols = NULL, facet_rows = NULL, plot_distribution = TRUE,
  print_plot = TRUE, save_plot = FALSE, return_plot = FALSE,
  out_dir = NULL, file_name = NULL, file_ext = "png", lab_size = 8,
  text_size = 6, outline_size = 0.7, fig_width = 8, fig_height = 9,
  fig_units = "cm", val_pal = c("black", "#050155", "#120172",
  "#3b008e", "#7200a9", "#8f00a0", "#ba187f", "#d9365b", "#ed5930",
  "#f76323", "#fa8600", "#f6a704", "#fad61e", "#fad61e"),
  fill_breaks = waiver(), patch_cols = c("mistyrose",
  "cornflowerblue"), patch_labs = c("Hot spots", "Cold spots"),
  val_lab = NULL)
}
\arguments{
\item{df}{A dataframe returned from \code{\link{get_patches}}.}

\item{patches}{A SpatialPolygonsDataFrame returned from \code{\link{get_patches}}.}

\item{bg_poly}{An optional background polygon. Can be a SpatialPolygonsDataFrame
or fortified dataframe (see \code{ggplot2::}\code{\link[ggplot2]{fortify}})
with the variables 'long', 'lat' and 'group'.}

\item{bg_colour}{Colour of the background polygon.}

\item{facet_cols}{Number of facet columns. Defaults to NULL.}

\item{facet_rows}{Number of facet rows. Defaults to NULL.}

\item{plot_distribution}{Should a histogram be plotted? Defaults to TRUE.}

\item{print_plot}{Should the resulting plots be printed? Defaults to FALSE.}

\item{save_plot}{Should the resulting plots be saved? Defaults to TRUE.}

\item{return_plot}{Should the resulting plots be returned? Defaults to FALSE.}

\item{out_dir}{Path to directory where plots should be saved (if applicable).}

\item{file_name}{Prefix for plot filenames (if applicable). If none
specified, uses generic names 'distribution' and 'patches'.}

\item{file_ext}{File extension. Defaults to '.png'.}

\item{lab_size}{Size of axes labels. Defaults to 8.}

\item{text_size}{Size of axes text and legend text. Defaults to 6.}

\item{outline_size}{Size of the outline. Defaults to 0.7.}

\item{fig_width}{Figure width (if saved). Defaults to 8.}

\item{fig_height}{Figure height (if saved). Defaults to 9.}

\item{fig_units}{Figure dimension units (if saved). Defaults to "cm".}

\item{val_pal}{Colour palette to use for raster. Defaults to palette
derived from a FLIR jpeg: \code{c("black", "#050155", "#120172", "#3b008e",
"#7200a9", "#8f00a0","#ba187f", "#d9365b", "#ed5930","#f76323", "#fa8600",
"#f6a704","#fad61e", "#fad61e")}.}

\item{fill_breaks}{Option to manually specify breaks in colourbar. Defaults
to \code{waiver()}, where breaks are computed by the transformation object
(see \code{ggplot2::}\code{scale_colour_gradient}).}

\item{patch_cols}{Colours for the patch borders (hot spot colour followed by
cold spot colour). Defaults to: \code{c("mistyrose", "cornflowerblue")}.}

\item{patch_labs}{Labels to use in patch outline legend. Defaults to 'Hot
spots' and 'Cold spots'.}

\item{val_lab}{Label to describe the variable of interest - corresponds to
the x axis of the histogram, and the fill legend of the raster plot.}

\item{facet}{Whether to plot facets by image ID. Defaults to FALSE.}
}
\description{
Plot hot and cold patches returned from \code{\link{get_patches}}.
}
\examples{
# FLIR temperature matrix ---------------------------------------------------
# Find hot and cold patches
flir_results <-
    get_patches(img = flir11835$flir_matrix,
                id = flir11835$photo_no)

# Look at the results for individual pixels
head(flir_results$df)

# Look at the patch statistics for hot and cold patches
flir_results$pstats

# Plot the patches
sp::plot(flir_results$patches)

# Plot using ThermStats::plot_patches
plot_patches(df = flir_results$df,
             patches = flir_results$patches,
             print_plot = TRUE,
             save_plot = FALSE)

# Worldclim2 temperature raster ---------------------------------------
# Dataset 'sulawesi_temp' represents mean January temperature for the
# island of Sulawesi

# Define projection and extent
img_proj <- raster::projection(sulawesi_temp)
img_extent <- raster::extent(sulawesi_temp)

# Find hot and cold patches
worldclim_results <-
 get_patches(img = sulawesi_temp,
             id = "sulawesi",
             style = "C",
             img_proj = img_proj,
             img_extent = img_extent)

# Look at the results for individual pixels
head(worldclim_results$df)

# Look at the patch statistics for hot and cold patches
worldclim_results$pstats

# Plot the patches
sp::plot(worldclim_results$patches)

# Plot using ThermStats::plot_patches
plot_patches(df = worldclim_results$df,
             patches = worldclim_results$patches,
             bg_poly = sulawesi_bg,
             print_plot = TRUE,
             save_plot = FALSE)

}
