% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHDI.R
\name{SHDI}
\alias{SHDI}
\title{SHDI}
\usage{
SHDI(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}
}
\description{
Helper function to calculate Shannon Diversity Index.
}
\examples{
SHDI(c(1:100, NA))
}
\keyword{internal}
