% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_by_group.R
\name{stats_by_group}
\alias{stats_by_group}
\title{stats_by_group}
\usage{
stats_by_group(metadata, mat_list, matrix_ID, grouping_var, round_val, ...)
}
\arguments{
\item{metadata}{A dataframe denoting the grouping of different matrices.}

\item{mat_list}{List of temperatures matrices.}

\item{matrix_ID}{Name of the metadata variable that identifies unique
temperature matrices. Should match element names in the list of matrices.}

\item{grouping_var}{The name of the metadata variable that denotes the grouping
of temperature matrices.}

\item{round_val}{Value to round to. Use 1 for no rounding.}

\item{...}{Use to specify statistics that should be calculated across all
pixels. Several helper functions are included for use here: perc_5, perc_95,
SHDI, SIDI. See examples below.}
}
\value{
A list containing:
 \item{raw_dat}{A list with one element per input thermal image. Each element is a numeric matrix of the
raw infrared data.}
 \item{camera_params}{A dataframe of callibration constants unique to each camera.}
}
\description{
Calculate thermal statistics across photos within groups.
}
\examples{
# Load raw data
raw_dat <- flir_raw$raw_dat
camera_params <- flir_raw$camera_params
metadata <- flir_raw$metadata

# Batch convert
mat_list <- batch_convert(raw_dat, write_results = FALSE)

# Calculate patch and pixel stats! ------------------------------------------

# Specifying mean, max and min for pixel stats
patch_stats_1 <- stats_by_group(metadata = metadata, mat_list = mat_list, matrix_ID = "photo_no", grouping_var = "rep_id", round_val = 0.5, mean, max, min)

# Specifying kurtosis and sknewness for pixel stats
library(moments)
patch_stats_2 <- stats_by_group(metadata = metadata, mat_list = mat_list, matrix_ID = "photo_no", grouping_var = "rep_id", round_val = 0.5, kurtosis, skewness)

# Specifying 5th and 95th percentiles for pixel stats
patch_stats_3 <- stats_by_group(metadata = metadata, mat_list = mat_list, matrix_ID = "photo_no", grouping_var = "rep_id", round_val = 0.5, perc_5, perc_95)

# Specifying Shannon and Simpson Diversity Indices for pixel stats
patch_stats_4 <- stats_by_group(metadata = metadata, mat_list = mat_list, matrix_ID = "photo_no", grouping_var = "rep_id", round_val = 0.5, SHDI, SIDI)
}
