% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_by_group.R
\name{stats_by_group}
\alias{stats_by_group}
\title{stats_by_group}
\usage{
stats_by_group(metadata, mat_list, matrix_id, style = "C", grouping_var,
  round_val, sum_stats = c("mean", "min", "max"))
}
\arguments{
\item{metadata}{A dataframe denoting the grouping of different matrices.}

\item{mat_list}{List of matrices.}

\item{matrix_id}{Name of the metadata variable that identifies unique
matrices. Should match element names in the list of matrices.}

\item{style}{Style to use when calculating neighbourhood weights using
\code{spdep::}\code{\link[spdep]{nb2listw}}.}

\item{grouping_var}{The name of the metadata variable that denotes the
grouping of matrices.}

\item{round_val}{Value to round to. Use 1 for no rounding.}

\item{k}{Number of neighbours to use when calculating nearest neighbours
using \code{spdep::}\code{\link[spdep]{knearneigh}}.}

\item{...}{Use to specify summary statistics that should be calculated across
all pixels. Several helper functions are included for use here:
\code{\link{perc_5}}, \code{\link{perc_95}},
\code{\link{SHDI}}, \code{\link{SIDI}},
\code{\link{kurtosis}} and \code{\link{skewness}}. Also see examples below.}
}
\value{
A dataframe with pixel and patch statistics. See
\code{\link{patch_stats}} for details of all the patch statistics returned.
}
\description{
Calculate summary and spatial statistics across multiple matrices within groups.
}
\examples{
# Load raw data
raw_dat <- flir_raw$raw_dat
camera_params <- flir_raw$camera_params
metadata <- flir_metadata

# Batch convert
mat_list <- batch_convert(raw_dat, write_results = FALSE)

# Calculate patch and pixel stats! ------------------------------------------

# Pixel stats = mean, max and min
patch_stats_1 <-
    stats_by_group(metadata = metadata,
                   mat_list = mat_list,
                   matrix_id = "photo_no",
                   grouping_var = "rep_id",
                   round_val = 0.5,
                   style = "C",
                   mean, max, min)

# Pixel stats = kurtosis and sknewness
patch_stats_2 <-
    stats_by_group(metadata = metadata,
                   mat_list = mat_list,
                   matrix_id = "photo_no",
                   grouping_var = "rep_id",
                   round_val = 0.5,
                   style = "C",
                   kurtosis, skewness)

# Pixel stats = 5th and 95th percentiles
patch_stats_3 <-
    stats_by_group(metadata = metadata,
                   mat_list = mat_list,
                   matrix_id = "photo_no",
                   grouping_var = "rep_id",
                   round_val = 0.5,
                   style = "C",
                   perc_5, perc_95)

# Pixel stats = Shannon and Simpson Diversity Indices
patch_stats_4 <-
    stats_by_group(metadata = metadata,
                   mat_list = mat_list,
                   matrix_id = "photo_no",
                   grouping_var = "rep_id",
                   round_val = 0.5,
                   style = "C",
                   SHDI, SIDI)
}
