% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_by_group.R
\name{stats_by_group}
\alias{stats_by_group}
\title{stats_by_group}
\usage{
stats_by_group(metadata, img_list, idvar, grouping_var = NULL,
  calc_connectivity = FALSE, conn_threshold = 1.5, patches = TRUE,
  style = "C", return_vals = c("df", "patches", "pstats"),
  round_val = NULL, sum_stats = c("mean", "max", "min"))
}
\arguments{
\item{metadata}{A dataframe denoting the grouping of different images.}

\item{img_list}{List or stack of numeric temperature matrices or rasters.}

\item{idvar}{Name of the metadata variable that identifies unique
images. Should match element names in the image list.}

\item{grouping_var}{The name of the metadata variable that denotes the
grouping of images.}

\item{style}{Style to use when calculating neighbourhood weights using
\code{spdep::}\code{\link[spdep]{nb2listw}}.}

\item{round_val}{Value to round to. Use 1 for no rounding.}

\item{sum_stats}{Summary statistics that should be calculated across
all pixels. Several helper functions are included for use here:
\code{\link{perc_5}}, \code{\link{perc_95}},
\code{\link{SHDI}}, \code{\link{SIDI}},
\code{\link{kurtosis}} and \code{\link{skewness}}. Also see examples below.}
}
\value{
A dataframe with pixel and patch statistics. See
\code{\link{patch_stats}} for details of all the patch statistics returned.
}
\description{
Calculate summary and spatial statistics across multiple images within groups.
}
\examples{
# Load raw data
raw_dat <- flir_raw$raw_dat
camera_params <- flir_raw$camera_params
metadata <- flir_metadata

# Batch convert
img_list <- batch_convert(raw_dat, write_results = FALSE)

# Calculate patch and pixel stats! ------------------------------------------

# Pixel stats = mean, max and min
patch_stats_1 <-
    stats_by_group(metadata = metadata,
                   img_list = img_list,
                   idvar = "photo_no",
                   style = "C",
                   grouping_var = "rep_id",
                   round_val = 0.5,
                   sum_stats = c("mean", "max", "min"))

# Pixel stats = kurtosis and sknewness
patch_stats_2 <-
    stats_by_group(metadata = metadata,
                   img_list = img_list,
                   idvar = "photo_no",
                   style = "C",
                   grouping_var = "rep_id",
                   round_val = 0.5,
                   sum_stats = c("kurtosis", "skewness"))

# Pixel stats = 5th and 95th percentiles
patch_stats_3 <-
    stats_by_group(metadata = metadata,
                   img_list = img_list,
                   idvar = "photo_no",
                   style = "C",
                   grouping_var = "rep_id",
                   round_val = 0.5,
                   sum_stats = c("perc_5", "perc_95"))

# Pixel stats = Shannon and Simpson Diversity Indices
patch_stats_4 <-
    stats_by_group(metadata = metadata,
                   img_list = img_list,
                   idvar = "photo_no",
                   style = "C",
                   grouping_var = "rep_id",
                   round_val = 0.5,
                   sum_stats = c("SHDI", "SIDI"))
}
